requirements.txt
# App
streamlit>=1.28.0,<1.40.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.6.0,<3.0.0
requests>=2.31.0,<3.0.0

# LLM
google-genai>=0.7.0,<1.0.0

# YouTube discovery and transcripts
yt-dlp>=2024.1.0
youtube-transcript-api>=0.6.0,<1.0.0

# Optional ASR
faster-whisper>=1.1.0,<2.0.0

# Optional official playlist publishing
google-auth>=2.30.0,<3.0.0
google-auth-oauthlib>=1.2.0,<2.0.0
google-api-python-client>=2.140.0,<3.0.0

# Testing
pytest>=8.0.0,<9.0.0
